When scoring weights are set at zero, the calculation isn't performed to increase speed
Added score tolerance to allow downhilling traversal similar to simulated annealing. It Seems to work better without reverting to the best key.

All the ciphers that it used to solve well, a revert never happened because it never got stuck at a local optima

Unfinished tabu list of peak keys, and logging of the plain text when the peak is hit.

Running key solver using text provided in keytext.txt file

Stat fixes